step: cpp now links to c code
The C++ (Vectorizer_GUI) target is now able to call code in the C (vec) target.
To make this easier a root CMakeLists.txt was added to make the whole repo one big cmake project with 2 targets (C and Cpp which links to the C stuff).
It is a little unnecessary to add dynamic library support, but I wanted to try anyway.